Correct degeneracy table
Previously degeneracy_table was incomplete, because it was using n_squared to generate the "primitive" vectors for generating group orbits, and that had a built-in assumption of O_h.

Eventually, the current implementation will be supplanted by one that doesn't need to be informed about the group---the group will be automatically determined by the specification of the volume and the boost.